+++ Source/Krypton Components/Krypton.Utilities/Components/Krypton AboutBox/General/Definitions.cs
@@ -1,0 +1,50 @@
+﻿#region BSD License
+/*
+ *
+ *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
+ *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2026 - 2026. All rights reserved.
+ *
+ */
+#endregion
+
+namespace Krypton.Utilities;
+
+#region AboutBox Definitions
+
+#region Enum AboutToolkitPage
+
+internal enum AboutToolkitPage
+{
+    GeneralInformation = 0,
+    Discord = 1,
+    DeveloperInformation = 2,
+    Versions = 3
+}
+
+#endregion
+
+#region Enum AboutBoxFileInformationPage
+
+public enum AboutBoxFileInformationPage
+{
+    Application = 0,
+    Assemblies = 1,
+    AssemblyDetails = 2
+}
+
+#endregion
+
+#region Enum AboutBoxPage
+
+public enum AboutBoxPage
+{
+    GeneralInformation = 0,
+    Description = 1,
+    FileInformation = 2,
+    Theme = 3,
+    ToolkitInformation = 4
+}
+
+#endregion
+
+#endregion